+++ R/05_LCA.R
@@ -22,6 +22,7 @@
 #'  \item{nobs}{Sample size (number of rows in the dataset).}
 #'  \item{Nclass}{Number of latent classes specified.}
 #'  \item{N_Cycle}{Number of EM algorithm iterations performed.}
+#'  \item{converge}{Logical value indicating whether the algorithm converged within maxiter iterations}
 #'  \item{TRP}{Test Reference Profile vector showing expected scores for each latent class.
 #'    Calculated as the column sum of the estimated class reference matrix.}
 #'  \item{LCD}{Latent Class Distribution vector showing the number of examinees assigned to each latent class.}
@@ -149,6 +150,7 @@ LCA <- function(U, ncls = 2, na = NULL, Z = NULL, w = NULL, maxiter = 100, verbo
     nobs = NROW(tmp$U),
     Nclass = ncls,
     N_Cycle = fit$iter,
+    converge = fit$converge,
     TRP = as.vector(TRP),
     LCD = as.vector(LCD),
     CMD = as.vector(CMD),