+++ R/16_OrdinalBiclustering.R
@@ -138,6 +138,7 @@ Biclustering.ordinal <- function(U,
     Fil <- create_filter_matrix(ncls)
   }
   # iteration start ---------------------------------------------------------
+  converge <- TRUE
   FLG <- TRUE
   while (FLG) {
     if (testell - oldtestell < 1e-8 * abs(oldtestell)) {
@@ -145,8 +146,10 @@ Biclustering.ordinal <- function(U,
       break
     }
     if (emt == maxemt) {
+      converge <- FALSE
       FLG <- FALSE
-      message("reach max iteration")
+      message("\nReached the maximum number of iterations.")
+      message("Warning: Algorithm may not have converged. Interpret results with caution.")
     }
 
     emt <- emt + 1
@@ -335,6 +338,7 @@ Biclustering.ordinal <- function(U,
     msg = msg,
     model = model,
     mic = mic,
+    converge = converge,
     nobs = nobs,
     Nclass = ncls,
     Nfield = nfld,